tests(graph): separate concerns into each unit test
