+++ docs/source/tests/test.rst
@@ -17,6 +17,5 @@ SASS
 Others
 ------
 
-.. automodule:: tests.test.test_graph
 .. automodule:: tests.test.test_half
 .. automodule:: tests.test.test_saxpy
