+++ tests/test/CMakeLists.txt
@@ -1,6 +1,5 @@
 add_pytest(environment)
 
-add_pytest(graph)
 add_pytest(half)
 add_pytest(saxpy)
